+++ docs/CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 -----
 - Add variant related objects
 - Add ``prediction_group`` to ``Resolwe`` object
+- Add ``PredictionPreset`` resource
 
 
 ===================